0.2 introduces role creation, removal, and addition during subscription for users. Still need to clean up rest of code base before merging to master.

+++ src/rust/src/bot/commands/unsubscribe.rs
@@ -14,21 +14,63 @@ use serenity:: {
 // unsubscribe the user calling this function to the match making group matching the name
 // of the channel that this function was called from
 pub async fn unsubscribe(context: &Context, message: &Message, _: Args) -> CommandResult {
-    // retrieve matchmaking group
     let reply;
+    // retrieve matchmaking group
     let group = match message.channel_id.name(&context).await {
         Some(g) => g,
         None => {
             reply = "error retrieving channel name.".to_string();
             let _ = message.channel_id.say(&context.http, &reply);
-            return Err(CommandError::from(reply));
+            return Err(CommandError::from(reply))
         }
     };
-    // execute database function
+    // retrieve guild
+    let guild = match message.guild(&context).await {
+        Some(g) => g,
+        None => {
+            reply = "error retrieving guild information.".to_string();
+            let _ = message.channel_id.say(&context.http, &reply);
+            return Err(CommandError::from(reply))
+        }
+    };
+    let guild = guild.read().await;
+    // retrieve member
+    let mut member = match guild.member(&context.http, &message.author.id).await {
+        Ok (m) => m,
+        Err(e) => {
+            reply = format!("error retrieving member: `{}`.", e);
+            let _ = message.channel_id.say(&context.http, &reply);
+            return Err(CommandError::from(reply))
+        }
+    };
+    // retrieve role
+    let role = match guild.role_by_name(&group) {
+        Some(r) => r,
+        None => {
+            reply = format!("error retrieving role `{}`.", &group);
+            let _ = message.channel_id.say(&context.http, &reply);
+            return Err(CommandError::from(reply))
+        }
+    };
+    // remove role from discord and database
     let result = {
+        // remove role from discord
+        if let Err(e) = member.remove_role(&context.http, role).await {
+            reply = format!("error removing role: `{}` for user: `{}` error: `{}`", role, message.author.name, e);
+            let _ = message.channel_id.say(&context.http, &reply);
+            return Err(CommandError::from(reply))            
+        }
         // retrieve database
-        let database_lock = context.data.read().await.get::<Database>().cloned().unwrap();
-        let database = database_lock.read().await;
+        let database = match context.data.read().await.get::<Database>().cloned() {
+            Some(d) => d,
+            None => {
+                reply = "error retrieving database object".to_string();
+                let _ = message.channel_id.say(&context.http, &reply);
+                return Err(CommandError::from(reply))                
+            }
+        };
+        // remove role from database
+        let database = database.read().await;
         match database.remove_mm_user(*message.author.id.as_u64(), &group).await {
             Ok (r) => r,
             Err(e) => {
@@ -38,7 +80,7 @@ pub async fn unsubscribe(context: &Context, message: &Message, _: Args) -> Comma
             }
         }
     };
-    // parse result
+    // return
     match result {
         0 => reply = format!("`{}` has been unsubscribed from the `{}` match making group.", message.author.name, group),
         1 => reply = format!("failed to add `{}` to the database.", message.author.name),
